Renamed shared_tasks to tasks
