+++ omigami/spec2vec/flows/training_flow.py
@@ -6,7 +6,7 @@ from omigami.config import IonModes, ION_MODES
 from omigami.gateways.data_gateway import InputDataGateway, SpectrumDataGateway
 from omigami.flow_config import FlowConfig
 
-from omigami.shared_tasks import DownloadData, DownloadParameters
+from omigami.tasks import DownloadData, DownloadParameters
 
 from omigami.spec2vec.tasks import (
     MakeEmbeddings,
