+++ omigami/tasks/__init__.py
@@ -1,0 +1,1 @@
+from omigami.tasks.download_data import DownloadData, DownloadParameters
