+++ omigami/test/shared_tasks/test_download_data.py
@@ -13,7 +13,7 @@ from omigami.config import SOURCE_URI_PARTIAL_GNPS
 from omigami.gateways.data_gateway import InputDataGateway
 from omigami.utils import create_prefect_result_from_path
 from omigami.gateways.input_data_gateway import FSInputDataGateway
-from omigami.shared_tasks.download_data import DownloadParameters, DownloadData
+from omigami.tasks.download_data import DownloadParameters, DownloadData
 from omigami.test.conftest import ASSETS_DIR
